// Pcf/DuplicateDetection/DuplicateDetectionPCF/enums/FieldTypes.ts
export enum FieldTypes {
  Currency = 'Currency',
  DateAndTime = 'DateAndTime.DateAndTime',
  DateOnly = 'DateAndTime.DateOnly',
  Decimal = 'Decimal',
  WholeNumber = 'Whole.None',
  Lookup = 'Lookup.Simple',
  MultiLineText = 'Multiple',
  OptionSet = 'OptionSet',
  SingleLineText = 'SingleLine.Text',
  TwoOptions = 'TwoOptions',
  OpenNestedButton = 'openNestedButton',
}